+++ src/main/java/org/example/service/PlanetCrudService.java
@@ -1,13 +1,7 @@
 package org.example.service;
 
-import org.example.DAO.ClientDao;
 import org.example.DAO.PlanetDao;
-import org.example.config.HibernateConfig;
-import org.example.models.ClientModel;
 import org.example.models.PlanetModel;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.Transaction;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -18,18 +12,10 @@ public class PlanetCrudService {
 
     public void savePlanet(PlanetModel planet){
 
-        Pattern pattern = Pattern.compile("\\W");
-        Matcher matcher = pattern.matcher(planet.getId());
-
-        if (matcher.find()) {
-            return;
-        } else if ((planet.getName().isEmpty())||(planet.getName().length()>500)){
-            return;
+        if (validateId(planet)){
+            planet.setId(planet.getId().toUpperCase());
+            planetDao.save(planet);
         }
-
-        planet.setId(planet.getId().toUpperCase());
-        planetDao.save(planet);
-
     }
 
     public PlanetModel findClientById(int id){
@@ -37,10 +23,29 @@ public class PlanetCrudService {
     }
 
     public void updateClient(PlanetModel planet){
-        planetDao.update(planet);
+        if(validateId(planet)){
+            planet.setId(planet.getId().toUpperCase());
+            planetDao.update(planet);
+        }
     }
 
     public void deleteClient(PlanetModel planet){
-        planetDao.delete(planet);
+        if(validateId(planet)){
+            planet.setId(planet.getId().toUpperCase());
+            planetDao.delete(planet);
+        }
+    }
+
+    private Boolean validateId(PlanetModel planet){
+        Pattern pattern = Pattern.compile("\\W");
+        Matcher matcher = pattern.matcher(planet.getId());
+        if (matcher.find()) {
+            System.out.println("the field 'ID' have incorrect symbol");
+            return false;
+        } else if ((planet.getName().isEmpty())||(planet.getName().length()>500)){
+            System.out.println("the field 'ID' incorrect");
+            return false;
+        }
+    return true;
     }
 }
